[Bronze II] Title: 숫자, Time: 8 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -10,10 +10,10 @@ int main()
     if(a==b) cout << 0 << "\n";
     else if(a>b){
         cout << a-b-1 << "\n";
-        for(int i=b+1;i<a;i++)  cout << i << " ";
+        for(long long i=b+1;i<a;i++)  cout << i << " ";
     }
     else{
         cout << b-a-1 << "\n";
-        for(int i=a+1;i<b;i++)  cout << i << " ";
+        for(long long int i=a+1;i<b;i++)  cout << i << " ";
     }
 }
